# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BalanceOut": {
        "description": "The customer's spendable balance.",
        "properties": {
          "balance_uzs": {
            "title": "Balance Uzs",
            "type": "string"
          }
        },
        "required": [
          "balance_uzs"
        ],
        "title": "BalanceOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MeOut": {
        "description": "The signed-in account as its owner sees it (the trade link is theirs to see).",
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "email_verified": {
            "title": "Email Verified",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "locale": {
            "enum": [
              "ru",
              "uz",
              "en"
            ],
            "title": "Locale",
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "steam_id": {
            "title": "Steam Id",
            "type": "string"
          },
          "trade_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Link"
          },
          "trade_link_checked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Link Checked At"
          },
          "trade_link_reason": {
            "anyOf": [
              {
                "enum": [
                  "invalid",
                  "private",
                  "trade_ban",
                  "hold",
                  "unavailable"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Link Reason"
          },
          "trade_link_verdict": {
            "anyOf": [
              {
                "enum": [
                  "ok",
                  "warn",
                  "bad"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Link Verdict"
          }
        },
        "required": [
          "id",
          "steam_id",
          "display_name",
          "avatar_url",
          "email",
          "email_verified",
          "locale",
          "trade_link",
          "trade_link_verdict",
          "trade_link_reason",
          "trade_link_checked_at",
          "roles",
          "created_at"
        ],
        "title": "MeOut",
        "type": "object"
      },
      "MePatchIn": {
        "additionalProperties": false,
        "description": "Editable profile fields; omitted = unchanged, ``email: null`` = clear.",
        "properties": {
          "email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "locale": {
            "anyOf": [
              {
                "enum": [
                  "ru",
                  "uz",
                  "en"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          }
        },
        "title": "MePatchIn",
        "type": "object"
      },
      "OrderCreateIn": {
        "additionalProperties": false,
        "description": "Buy one skin: the offer the buyer chose and the price the panel showed them.",
        "properties": {
          "listing_id": {
            "exclusiveMinimum": 0.0,
            "title": "Listing Id",
            "type": "integer"
          },
          "price_uzs": {
            "exclusiveMinimum": 0.0,
            "title": "Price Uzs",
            "type": "integer"
          },
          "slug": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "listing_id",
          "price_uzs"
        ],
        "title": "OrderCreateIn",
        "type": "object"
      },
      "OrderOut": {
        "description": "An order as its owner sees it.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "delivered_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivered At"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "number": {
            "title": "Number",
            "type": "string"
          },
          "paid_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paid At"
          },
          "paid_with": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paid With"
          },
          "payable": {
            "title": "Payable",
            "type": "boolean"
          },
          "phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase"
          },
          "price_usd": {
            "title": "Price Usd",
            "type": "string"
          },
          "price_uzs": {
            "title": "Price Uzs",
            "type": "string"
          },
          "refunded_to": {
            "anyOf": [
              {
                "const": "balance",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refunded To"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "paid",
              "buying",
              "trade_sent",
              "delivered",
              "cancelled",
              "failed",
              "returned"
            ],
            "title": "Status",
            "type": "string"
          },
          "trade": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SkinTradeOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "number",
          "status",
          "slug",
          "name",
          "phase",
          "image_url",
          "price_uzs",
          "price_usd",
          "created_at",
          "expires_at",
          "paid_at",
          "delivered_at",
          "paid_with",
          "refunded_to",
          "payable",
          "trade"
        ],
        "title": "OrderOut",
        "type": "object"
      },
      "OrdersPage": {
        "description": "A page of the customer's orders, newest first.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "OrdersPage",
        "type": "object"
      },
...
      "TopupIn": {
        "additionalProperties": false,
        "description": "Open a balance top-up.",
        "properties": {
          "amount_uzs": {
            "title": "Amount Uzs",
            "type": "integer"
          },
          "locale": {
            "enum": [
              "ru",
              "uz",
              "en"
            ],
            "title": "Locale",
            "type": "string"
          },
          "provider": {
            "maxLength": 16,
            "minLength": 1,
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "amount_uzs",
          "provider",
          "locale"
        ],
        "title": "TopupIn",
        "type": "object"
      },
      "TopupOut": {
        "description": "A top-up as its owner sees it.",
        "properties": {
          "amount_uzs": {
            "title": "Amount Uzs",
            "type": "string"
          },
          "awaiting_kassa": {
            "title": "Awaiting Kassa",
            "type": "boolean"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "intent_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent Url"
          },
          "number": {
            "title": "Number",
            "type": "string"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "status": {
            "enum": [
              "pending",
              "succeeded",
              "expired",
              "reversed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "number",
          "amount_uzs",
          "provider",
          "status",
          "expires_at",
          "intent_url",
          "awaiting_kassa"
        ],
        "title": "TopupOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "csmarket API",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "description": "Profile, roles and trade-link state.",
        "operationId": "get_me_api_v1_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The signed-in account",
        "tags": [
          "me"
        ]
      },
      "patch": {
        "description": "Partial update; replays a stored response for a repeated ``Idempotency-Key``.",
        "operationId": "patch_me_api_v1_me_patch",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MePatchIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edit locale or email",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/orders": {
      "get": {
        "description": "My orders, newest first, 20 a page; cancelled and expired unpaid ones are left out.",
        "operationId": "get_my_orders_api_v1_me_orders_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "My orders",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/api/v1/orders": {
      "post": {
        "description": "Open an order for the chosen offer at the price the panel showed.\n\n``Idempotency-Key`` (16..160 chars) is required. A replayed key answers 200 with the\nstored order, whatever the body says. 409 ``code``s: ``buying_disabled``,\n``trade_link_missing``, ``trade_link_bad`` (+ ``reason``), ``price_changed`` (+ the new\n``price_uzs``), ``offer_gone`` (+ ``next_offer`` or ``null``). 404 for an unknown item;\n503 ``rate_unavailable`` without a fresh so\u02bbm rate.",
        "operationId": "post_order_api_v1_orders_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Buy one skin",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/v1/orders/{number}": {
      "get": {
        "description": "The owner's order; anyone else's, an unknown or a malformed number is a 404.",
        "operationId": "get_order_api_v1_orders__number__get",
        "parameters": [
          {
            "in": "path",
            "name": "number",
            "required": true,
            "schema": {
              "title": "Number",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One of my orders",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/api/v1/wallet": {
      "get": {
        "description": "Spendable so\u02bbm; ``\"0\"`` before the first top-up.",
        "operationId": "get_balance_api_v1_wallet_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "My balance",
        "tags": [
          "wallet"
        ]
      }
    },
...
    "/api/v1/wallet/topups": {
      "post": {
        "description": "Open a top-up in the chosen kassa; the response says where to pay.\n\n``Idempotency-Key`` is required. The same key with the same amount and kassa returns\nthe same top-up (201 again); with another amount or kassa it is a 409\n``idempotency_mismatch``.",
        "operationId": "post_topup_api_v1_wallet_topups_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopupIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopupOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Open a balance top-up",
        "tags": [
          "wallet"
        ]
      }
    },
    "/api/v1/wallet/topups/{number}": {
      "get": {
        "description": "The owner's top-up; anyone else's (or an unknown number) is a 404, never a 403.",
        "operationId": "get_topup_api_v1_wallet_topups__number__get",
        "parameters": [
          {
            "in": "path",
            "name": "number",
            "required": true,
            "schema": {
              "title": "Number",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "ru",
              "enum": [
                "ru",
                "uz",
                "en"
              ],
              "title": "Locale",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopupOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One of my top-ups",
        "tags": [
          "wallet"
        ]
      }
    },
...
  }
}